+++ packages/theme/src/semantic-color-contrast-pairs.ts
@@ -16,6 +16,18 @@ export const SEMANTIC_COLOR_CONTRAST_PAIRS = [
 		foreground: 'foreground.interactive.neutral-weak',
 	},
 	{
+		background: 'background.interactive.neutral-active',
+		foreground: 'foreground.interactive.neutral',
+	},
+	{
+		background: 'background.interactive.neutral-active',
+		foreground: 'foreground.interactive.neutral-weak',
+	},
+	{
+		background: 'background.interactive.neutral-weak-active',
+		foreground: 'foreground.interactive.neutral-active',
+	},
+	{
 		background: 'background.surface.neutral',
 		foreground: 'foreground.content.neutral',
 	},